+++ app/cdap/components/ColumnViewPanel/components/SelectColumnsList/index.tsx
@@ -20,40 +20,16 @@ import NoRecordScreen from 'components/NoRecordScreen';
 import T from 'i18n-react';
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
+import {
+  ISelectColumnsListProps,
+  IDataQuality,
+  IDataQualityRecord,
+} from 'components/ColumnViewPanel/components/SelectColumnsList/types';
 
 const COLUMNS = T.translate('features.WranglerNewUI.ColumnViewPanel.columns');
 const NULL_VALUES = T.translate('features.WranglerNewUI.ColumnViewPanel.nullValues');
 
 import { IHeaderNamesList } from 'components/GridTable/types';
-
-export interface ISelectColumnsListProps {
-  columnData: IHeaderNamesList[];
-  dataQuality: IDataQuality;
-  searchTerm: string;
-  setColumnSelected: (columnName: string) => void;
-  onColumnSelection: (column: string) => void;
-  selectedColumn: string;
-  handleCoumnUnSelect: () => void;
-}
-
-export interface IDataQuality {
-  [key: string]: unknown;
-}
-
-export interface IDataQualityRecord {
-  label: string;
-  value: number;
-}
-
-export interface ISelectColumnsTableRowProps {
-  eachFilteredColumn: IHeaderNamesList;
-  filteredColumnIndex: number;
-  dataQualityList: IDataQualityRecord[];
-  setColumnSelected: (columnName: string) => void;
-  onColumnSelection: (column: string) => void;
-  selectedColumn: string;
-  handleCoumnUnSelect: () => void;
-}
 
 const CustomTableContainer = styled(TableContainer)`
   &.MuiTableContainer-root {
@@ -136,7 +112,7 @@ const getTableBody = (
   setColumnSelected,
   onColumnSelection,
   selectedColumn,
-  handleCoumnUnSelect
+  handleCoumnDeSelect
 ) => {
   {
     return filteredColumns.length !== 0 ? (
@@ -149,7 +125,7 @@ const getTableBody = (
             setColumnSelected={setColumnSelected}
             onColumnSelection={onColumnSelection}
             selectedColumn={selectedColumn}
-            handleCoumnUnSelect={handleCoumnUnSelect}
+            handleCoumnDeSelect={handleCoumnDeSelect}
           />
         );
       })
@@ -169,7 +145,7 @@ export default function({
   setColumnSelected,
   onColumnSelection,
   selectedColumn,
-  handleCoumnUnSelect,
+  handleCoumnDeSelect,
 }: ISelectColumnsListProps) {
   const [filteredColumns, setFilteredColumns] = useState<IHeaderNamesList[]>(columnData);
   const [dataQualityList, setDataQualityList] = useState<IDataQualityRecord[]>([]);
@@ -221,7 +197,7 @@ export default function({
             setColumnSelected,
             onColumnSelection,
             selectedColumn,
-            handleCoumnUnSelect
+            handleCoumnDeSelect
           )}
         </Table>
       </CustomTableContainer>
