+++ app/cdap/components/ColumnViewPanel/components/SelectColumnsList/types.ts
@@ -1,0 +1,46 @@
+/*
+ * Copyright © 2022 Cask Data, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not
+ * use this file except in compliance with the License. You may obtain a copy of
+ * the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations under
+ * the License.
+ */
+
+import { IHeaderNamesList } from 'components/GridTable/types';
+
+export interface ISelectColumnsListProps {
+  columnData: IHeaderNamesList[];
+  dataQuality: IDataQuality;
+  searchTerm: string;
+  setColumnSelected: (columnName: string) => void;
+  onColumnSelection: (column: string) => void;
+  selectedColumn: string;
+  handleCoumnDeSelect: () => void;
+}
+
+export interface IDataQuality {
+  [key: string]: unknown;
+}
+
+export interface IDataQualityRecord {
+  label: string;
+  value: number;
+}
+
+export interface ISelectColumnsTableRowProps {
+  eachFilteredColumn: IHeaderNamesList;
+  filteredColumnIndex: number;
+  dataQualityList: IDataQualityRecord[];
+  setColumnSelected: (columnName: string) => void;
+  onColumnSelection: (column: string) => void;
+  selectedColumn: string;
+  handleCoumnDeSelect: () => void;
+}
