+++ app/cdap/components/GridTable/types.ts
@@ -34,7 +34,6 @@ interface ISummary {
 export interface IExecuteAPIResponse {
   headers: string[];
   types: IRecords;
-  // values: IRecords[];
   values: IValues[];
   summary: ISummary;
   message: string;
@@ -43,7 +42,6 @@ export interface IExecuteAPIResponse {
 export interface IHeaderNamesList {
   name: string;
   label: string;
-  // type: Array<string | boolean | Record<string, IType>>;
   type: string[];
 }
